Extract version embedding to a helper
