+++ core/trino-main/src/test/java/io/trino/util/TestDecoratingListeningExecutorService.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.dispatcher;
+package io.trino.util;
 
 import com.google.common.util.concurrent.ListeningExecutorService;
 import org.junit.jupiter.api.Test;
